requirements.txt
flask==3.0.3
itsdangerous==2.2.0
jinja2==3.1.4
werkzeug==3.0.3
python-dateutil==2.9.0.post0
bcrypt==4.1.3
cloudinary==1.41.0
requests==2.32.3
gunicorn==22.0.0
